rename to TransferReceipt
