+++ connect/src/protocols/gatewayTransfer.ts
@@ -28,11 +28,12 @@ import {
   nativeChainAddress,
   toGatewayMsg,
   toNative,
+  AttestationId,
 } from "@wormhole-foundation/sdk-definitions";
 import { signSendWait } from "../common";
 import { fetchIbcXfer, isTokenBridgeVaaRedeemed, retry } from "../tasks";
 import { Wormhole } from "../wormhole";
-import { AttestationId, TransferState, WormholeTransfer } from "../wormholeTransfer";
+import { TransferState, WormholeTransfer } from "../wormholeTransfer";
 
 type GatewayContext<N extends Network> = ChainContext<
   N,
